Move all download code into 01_install_requirement.sh
New ansible playbook to download packages 


Install redhat-lsb-core  using yum to make  `ansible_lsb` available

Fix typos


Update package playbook with two level of centos: rhel7 and rhel8


Remove some packages that are  not in RHEL8

+++ host_cleanup.sh
@@ -42,14 +42,16 @@ ANSIBLE_FORCE_COLOR=true ansible-playbook \
     -i vm-setup/inventory.ini \
     -b -vvv vm-setup/teardown-playbook.yml
 
-sudo rm -rf /etc/NetworkManager/conf.d/dnsmasq.conf
 # There was a bug in this file, it may need to be recreated.
-if [ "$MANAGE_PRO_BRIDGE" == "y" ]; then
-    sudo ifdown provisioning || true
-    sudo rm -f /etc/sysconfig/network-scripts/ifcfg-provisioning || true
-fi
-# Leaving this around causes issues when the host is rebooted
-if [ "$MANAGE_BR_BRIDGE" == "y" ]; then
-    sudo ifdown baremetal || true
-    sudo rm -f /etc/sysconfig/network-scripts/ifcfg-baremetal || true
+if [[ $OS == "centos" ]]; then
+  sudo rm -rf /etc/NetworkManager/conf.d/dnsmasq.conf
+  if [ "$MANAGE_PRO_BRIDGE" == "y" ]; then
+      sudo ifdown provisioning || true
+      sudo rm -f /etc/sysconfig/network-scripts/ifcfg-provisioning || true
+  fi
+  # Leaving this around causes issues when the host is rebooted
+  if [ "$MANAGE_BR_BRIDGE" == "y" ]; then
+      sudo ifdown baremetal || true
+      sudo rm -f /etc/sysconfig/network-scripts/ifcfg-baremetal || true
+  fi
 fi
